updating scores
